Extract the video reading logic from the mp4 component
